Move getToken method to utils
